:recycle: tree reorganization
