--- ozkour-back/config/routes/error.js
@@ -1,9 +1,0 @@
-module.exports = [
-  {
-    method: '*',
-    path: '/{any*}',
-    handler: function (request, h) {
-      return '404 Error! Page Not Found!'
-    }
-  }
-]
