--- ozkour-back/config/routes/test-route.js
@@ -1,9 +1,0 @@
-module.exports = [
-  {
-    method: 'GET',
-    path: '/ping',
-    handler: function (request, h) {
-      return h.response('OK!').code(200)
-    }
-  }
-]
