--- ozkour-back/domain/type/drive.ts
@@ -1,20 +1,0 @@
-export type Drive = {
-    getIdOfTalkFileByYear : getIdOfTalkFileByYear,
-    copyDocument : copyDocument
-}
-
-enum folder {
-    "emailing"
-}
-
-
-type getIdOfTalkFileByYear = (
-    year: string,
-) => Promise<string>
-
-type copyDocument = (
-    folder: folder,
-    name:string
-) => Promise<string>
-
-
