--- ozkour-back/index.js
@@ -1,9 +1,0 @@
-'use strict'
-const server = require('./server')
-
-const start = async () => {
-  await server.init()
-  await server.start()
-}
-
-start()
